Refactor
to avoid "invalidProtocolError" from appearing in API documentation as
one of the "Models"

/components/responses don't appear to be usable in oneOf collections,
and it appears necessary to define one response with a oneOf collection
of schemas. Defining a $ref for this schema makes it appear in the API's
list of models, which are currently only populated with User, Event,
Tag, etc. I'm choosing to duplicate code instead of awkwardly adding
invalidProtocolError to that list.