Add docs for /pages endpoints and switch from JSON to YAML
The SwaggerHub API IDE
(https://app.swaggerhub.com/apis/Ball-State-CBER/muncie-events/1.0.0)
takes input as YAML. Also saving it as YAML in this codebase makes more
sense than continuing to convert it to JSON to commit changes in this
codebase. Actually, I'm not sure why I was originally doing that.
Something may have given me the impression that the swagger-ui.js code
could only take JSON input, but that's not actually the case.