Adjust how "pages" work on event index
Before, pages were broken up by sequential months, and if there is a
gap of an empty month before the next event, that next event won't be
shown. This commit changes pages to have up to 30 events, plus however
many events are necessary to display all events on the last date.